requirements.txt
google-adk>=0.1.0
google-genai>=0.3.0
asyncio
typing-extensions>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.3.0
